markers + event
added markers for all hotels in a city, added a event to observe a click on the marker that gets the long and latitude. this can be used for retrieving specific hotel information from the dataframe but still has to be implemented. planning to show reviews in the side bar under the filters